+++ src/utils/vehicleUtil.ts
